chore: add document service and organize project structure
