--- src/controller/dashboard.ts
@@ -1,10 +1,0 @@
-import { factory } from '../factory';
-import { authMiddleware } from '../hooks/server/auth';
-
-const dashboardApp = factory.createApp();
-
-dashboardApp.get('/', authMiddleware, (c) => {
-	return c.html('');
-});
-
-export default dashboardApp;
