+++ src/factory.ts
@@ -1,6 +1,7 @@
 import { createFactory } from 'hono/factory';
 import type { Variables } from '../server';
 import { UserService } from './lib/server/user/service/user.service';
+import { DocumentService } from './lib/server/document/service/document.service';
 
 export const factory = createFactory<{
 	Variables: Variables;
@@ -19,5 +20,9 @@ export class ServiceFactory {
 
 	static getUserService(): UserService {
 		return ServiceFactory.getService(UserService);
+	}
+
+	static getDocumentService() {
+		return ServiceFactory.getService(DocumentService);
 	}
 }
